+++ src/main/kotlin/com/gemini/krakenbot/view/component/DashboardShellComponent.kt
@@ -2,10 +2,12 @@ package com.gemini.krakenbot.view.component
 
 import com.gemini.krakenbot.view.util.CssClass
 import com.gemini.krakenbot.view.util.HtmxAttrs
+import com.gemini.krakenbot.view.util.HtmxValues
 import com.gemini.krakenbot.view.util.Routes
 import com.gemini.krakenbot.view.util.ViewText
 import com.gemini.krakenbot.view.util.commonMetadataAndStyles
 import com.gemini.krakenbot.view.util.div
+import com.gemini.krakenbot.view.util.rebalancerJsSrc
 import kotlinx.html.*
 
 class DashboardShellComponent {
@@ -21,12 +23,12 @@ class DashboardShellComponent {
         html.body {
             div(CssClass.Layout.Container) {
                 div {
-                    attributes[HtmxAttrs.HX_EXT] = "sse"
+                    attributes[HtmxAttrs.HX_EXT] = HtmxValues.EXT_SSE
                     attributes[HtmxAttrs.SSE_CONNECT] = Routes.API_STATUS_STREAM
 
                     div {
                         attributes[HtmxAttrs.HX_GET] = Routes.FRAGMENT_DASHBOARD
-                        attributes[HtmxAttrs.HX_TRIGGER] = "load, sse:message"
+                        attributes[HtmxAttrs.HX_TRIGGER] = HtmxValues.TRIGGER_LOAD_SSE_MESSAGE
 
                         div(CssClass.Loading.SpinnerContainer) {
                             div(CssClass.Loading.Spinner) {}
@@ -35,7 +37,7 @@ class DashboardShellComponent {
                     }
                 }
             }
-            script(src = Routes.STATIC_REBALANCER_JS) {}
+            script(src = rebalancerJsSrc()) {}
         }
     }
 }